refactor(select): split monolithic select.rs into modules
Split 748-line select.rs into focused modules following the combobox pattern:
- mod.rs: struct Select + builders + pub API (~190 lines)
- render.rs: impl View for Select (~170 lines)
- input.rs: impl Interactive for Select (~110 lines)
- filter.rs: fuzzy search and filtering (~70 lines)
- state.rs: open/close/toggle + selection helpers (~160 lines)

All 246 select-related tests continue to pass.
